moved some mocked data from MovieForm helper to MockData.ts, created MockedDataTypes file
