+++ src/components/MovieForm/types.ts
@@ -1,13 +1,3 @@
 export interface IMovieFormProps {
   type: 'add' | 'edit' | 'delete';
 }
-
-export interface IMovieFormFields {
-  addAndEdit: {
-    label: string;
-    type: 'text' | 'textarea' | 'select' | 'date';
-    placeholder: string;
-    options?: string[];
-    id?: string;
-  }[];
-}
