+++ src/data/MockData.ts
@@ -1,16 +1,19 @@
 import * as Movie1 from 'Images/movie-1.png';
 import * as Movie2 from 'Images/movie-2.png';
 import * as Movie3 from 'Images/movie-3.png';
+import {
+  IMoviesListData,
+  IMovieFormFields,
+  IGenreFilterList,
+} from './MockedDataTypes';
 
-const genreFilterList = ['All', 'Documentary', 'Comedy', 'Horror', 'Crime'];
-
-export interface IMoviesListData {
-  name: string;
-  genre: string;
-  year: number;
-  image: any;
-  id: number;
-}
+const genreFilterList: IGenreFilterList = [
+  'All',
+  'Documentary',
+  'Comedy',
+  'Horror',
+  'Crime',
+];
 
 const moviesList: IMoviesListData[] = [
   {
@@ -36,4 +39,44 @@ const moviesList: IMoviesListData[] = [
   },
 ];
 
-export { moviesList, genreFilterList };
+const movieFormFields: IMovieFormFields = {
+  addAndEdit: [
+    {
+      label: 'Title',
+      type: 'text',
+      placeholder: 'Movie Title',
+    },
+    {
+      label: 'Release Date',
+      type: 'date',
+      placeholder: 'Select Date',
+    },
+    {
+      label: 'Movie Url',
+      type: 'text',
+      placeholder: 'https://',
+    },
+    {
+      label: 'Rating',
+      type: 'text',
+      placeholder: '7.8',
+    },
+    {
+      label: 'Genre',
+      type: 'select',
+      placeholder: 'Select Genre',
+    },
+    {
+      label: 'Runtime',
+      type: 'text',
+      placeholder: 'minutes',
+    },
+    {
+      label: 'Overview',
+      type: 'textarea',
+      placeholder: 'Movie Description',
+    },
+  ],
+};
+
+export { moviesList, genreFilterList, movieFormFields };
